update
* removed unused code
* fixed a minor bug of class embedding in DiT
* added dataset split function
* improved GPU memory usage efficiency while loading model

+++ akane2/module/diffusion.py
@@ -50,7 +50,7 @@ class ClassLabelEmbed(nn.Module):
         :param label: label tokens;  shape: (n_b, 1)
         :return: embedded labels;    shape: (n_b, 1, n_f)
         """
-        return self.embed(label)
+        return self.embed(label.to(torch.long))
 
 
 class TextLabelEmbed(nn.Module):
